* Refactoring * new Trait WaitFor * WaitForActions mixing in WaitFor trait
