+++ code/Server/src/main/scala/net/selenate/server/sessions/actions/WaitForAction.scala
@@ -11,44 +11,11 @@ import org.openqa.selenium.firefox.FirefoxDriver
 import org.openqa.selenium.{ By, WebElement }
 
 import scala.collection.JavaConversions._
-import scala.annotation.tailrec
-
-object WaitForAction {
-  private val timeout    = 20000
-  private val resolution = 250
-
-  def waitForPredicate[T](predicate: => Option[T]): Option[T] = {
-    @tailrec
-    def waitForDoit(end: Long, resolution: Long, predicate: => Option[T]): Option[T] = {
-      val current = System.currentTimeMillis
-      val remaining = end - current
-
-      if (remaining < 0) {
-        None  // Timeout
-      } else {
-        val p = predicate
-        if (p.isDefined) {
-          p  // Predicate evaluated to true
-        } else {
-          // Do not oversleep.
-          val sleep = scala.math.min(resolution, remaining)
-          Thread.sleep(sleep)
-          waitForDoit(end, resolution, predicate)
-        }
-      }
-    }
-
-    val end = System.currentTimeMillis + timeout
-    waitForDoit(end, resolution, predicate)
-  }
-}
-
 
 class WaitForAction(val d: FirefoxDriver)
     extends IAction[SeReqWaitFor, SeResWaitFor]
-    with ActionCommons {
-  import WaitForAction._
-
+    with ActionCommons
+    with WaitFor {
   protected val log = Log(classOf[WaitForAction])
 
   def act = { arg =>
