+++ code/Server/src/main/scala/net/selenate/server/sessions/actions/WaitForBrowserPageAction.scala
@@ -9,45 +9,12 @@ import req._
 import org.openqa.selenium.firefox.FirefoxDriver
 import org.openqa.selenium.{ By, WebElement }
 import scala.collection.JavaConversions._
-import scala.annotation.tailrec
 import net.selenate.common.user.BrowserPage
-
-object WaitForBrowserPageAction {
-  private val timeout    = 20000
-  private val resolution = 250
-
-  def waitForPredicate[T](predicate: => Option[T]): Option[T] = {
-    @tailrec
-    def waitForDoit(end: Long, resolution: Long, predicate: => Option[T]): Option[T] = {
-      val current = System.currentTimeMillis
-      val remaining = end - current
-
-      if (remaining < 0) {
-        None  // Timeout
-      } else {
-        val p = predicate
-        if (p.isDefined) {
-          p  // Predicate evaluated to true
-        } else {
-          // Do not oversleep.
-          val sleep = scala.math.min(resolution, remaining)
-          Thread.sleep(sleep)
-          waitForDoit(end, resolution, predicate)
-        }
-      }
-    }
-
-    val end = System.currentTimeMillis + timeout
-    waitForDoit(end, resolution, predicate)
-  }
-}
-
 
 class WaitForBrowserPageAction(val d: FirefoxDriver)
     extends IAction[SeReqWaitForBrowserPage, SeResWaitForBrowserPage]
-    with ActionCommons {
-  import WaitForBrowserPageAction._
-
+    with ActionCommons
+    with WaitFor {
   protected val log = Log(classOf[WaitForBrowserPageAction])
 
   def act = { arg =>
@@ -58,6 +25,7 @@ class WaitForBrowserPageAction(val d: FirefoxDriver)
       case _       => null;
     }
 
+    // TODO SeResWaitForBrowserPage BasePage => SePage and conversion somewhere else
     new SeResWaitForBrowserPage(res.isDefined, ret)
   }
